requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Scheduler
apscheduler==3.10.4

# Environment Variables
python-dotenv==1.0.0

# Security & Auth
pyjwt==2.8.0
passlib[bcrypt]==1.7.4

# Email
sendgrid==6.11.0

# SMS
twilio==8.11.1

# QR Code Generation
qrcode[pil]==7.4.2

# Utilities
python-multipart==0.0.6

# LLM Integration (Stage 5)
langchain==0.1.0
langchain-openai==0.0.2
openai==1.7.2

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0

# Development
black==24.1.1
flake8==7.0.0
